+++ e2e/e2etests/test_zevm_to_evm_call_revert.go
@@ -5,28 +5,27 @@ import (
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/stretchr/testify/require"
-	"github.com/zeta-chain/protocol-contracts/pkg/gatewayevm.sol"
+	"github.com/zeta-chain/protocol-contracts/pkg/gatewayzevm.sol"
 
 	"github.com/zeta-chain/node/e2e/runner"
 	"github.com/zeta-chain/node/e2e/utils"
+	"github.com/zeta-chain/node/testutil/sample"
 	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 )
 
-func TestETHDepositAndCallRevertWithCall(r *runner.E2ERunner, args []string) {
-	require.Len(r, args, 1)
-
-	amount := utils.ParseBigInt(r, args[0])
-
-	r.ApproveERC20OnEVM(r.GatewayEVMAddr)
+func TestZEVMToEVMCallRevert(r *runner.E2ERunner, args []string) {
+	require.Len(r, args, 0)
 
 	payload := randomPayload(r)
 
-	r.AssertTestDAppEVMCalled(false, payload, amount)
+	r.ApproveETHZRC20(r.GatewayZEVMAddr)
 
-	// perform the deposit
-	tx := r.ETHDepositAndCall(r.TestDAppV2ZEVMAddr, amount, []byte("revert"),
-		gatewayevm.RevertOptions{
-			RevertAddress:    r.TestDAppV2EVMAddr,
+	// perform the withdraw
+	tx := r.ZEVMToEMVCall(
+		sample.EthAddress(), // non-existing address
+		[]byte("revert"),
+		gatewayzevm.RevertOptions{
+			RevertAddress:    r.TestDAppV2ZEVMAddr,
 			CallOnRevert:     true,
 			RevertMessage:    []byte(payload),
 			OnRevertGasLimit: big.NewInt(200000),
@@ -35,17 +34,16 @@ func TestETHDepositAndCallRevertWithCall(r *runner.E2ERunner, args []string) {
 
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
-	r.Logger.CCTX(*cctx, "deposit")
+	r.Logger.CCTX(*cctx, "call")
 	require.Equal(r, crosschaintypes.CctxStatus_Reverted, cctx.CctxStatus.Status)
 
-	// check the payload was received on the contract
-	r.AssertTestDAppEVMCalled(true, payload, big.NewInt(0))
+	r.AssertTestDAppZEVMCalled(true, payload, big.NewInt(0))
 
 	// check expected sender was used
-	senderForMsg, err := r.TestDAppV2EVM.SenderWithMessage(
+	senderForMsg, err := r.TestDAppV2ZEVM.SenderWithMessage(
 		&bind.CallOpts{},
 		[]byte(payload),
 	)
 	require.NoError(r, err)
-	require.Equal(r, r.EVMAuth.From, senderForMsg)
+	require.Equal(r, r.ZEVMAuth.From, senderForMsg)
 }