+++ changelog.md
@@ -15,6 +15,8 @@
 * [3534] (https://github.com/zeta-chain/node/pull/3534) - Add Sui deposit & depositAndCall
 * [3541](https://github.com/zeta-chain/node/pull/3541) - implement `MsgUpdateZRC20Name` to update the name or symbol of a ZRC20 token
 * [3520](https://github.com/zeta-chain/node/pull/3520) - SPL withdraw and call integration
+* [3439](https://github.com/zeta-chain/node/pull/3439) - use protocol contracts V2 with TON deposits
+* [3522](https://github.com/zeta-chain/node/pull/3522) - add `MsgDisableFastConfirmation` to disable fast confirmation. This message can be triggered by the emergency policy.
 
 ### Refactor
 
@@ -31,6 +33,23 @@
 
 * [3430](https://github.com/zeta-chain/node/pull/3430) - add simulation test for MsgWithDrawEmission
 * [3503](https://github.com/zeta-chain/node/pull/3503) - add check in e2e test to ensure deletion of stale ballots
+
+## v28.0.0
+
+v28 is based on the release/v27 branch rather than develop
+
+### Fixes
+* [3563](https://github.com/zeta-chain/node/pull/3563) - upgrade cosmos-sdk to v0.50.12 to resolve GHSA-x5vx-95h7-rv4p
+
+## v27.0.5
+
+### Fixes
+* [3554](https://github.com/zeta-chain/node/pull/3554) - disable observation of direct to TSS Address deposits on Arbitrum and Avalanch networks
+
+## v27.0.4
+
+### Fixes
+* [3508](https://github.com/zeta-chain/node/pull/3508) - fix empty from field in eth receipt rpc method
 
 ## v27.0.1